All proyect was ordered. All modules were redistributed And the test were divided in modules.
